Database Updates
Many class files were changed because we added a .jar file to the classpath. The big changes are that the GameData database was just created, so we need to add that to the project files. The second big change is that now the MapCreationManager will create a connection to the database. We also commented out the outputs of map creation manager just so we can see important print statements related to the database without having it cluttered by the maps.